goapi: allow inclusion of GOPATH
The rigor of maintaining api is a goal for all projects, not just the
golang stdlib. This tool is useful to the broader audience. This
`-gopath` flag enables the specifies package name arguments to be in
GOPATH. In this way projects can validate their API compatibility.

Forking suggested by bradfitz https://go-review.googlesource.com/#/c/15140/1